refactor(ca): add some changes in database package
