+++ src/pkg/database/model/wallet.ts
@@ -1,5 +1,5 @@
 import { Entity, PrimaryGeneratedColumn, Column, CreateDateColumn } from "typeorm";
-import { Wallet } from "../../../sychronizer";
+import { Network, Wallet } from "../../../synchronizer";
 
 @Entity({
   name: "wallet",
@@ -9,13 +9,17 @@ export class WalletModel {
   id: string;
 
   @Column()
-  address: string; // DISTINC
+  network: Network;
+
+  @Column()
+  address: string;
 
   @CreateDateColumn()
   createdAt: Date;
 
   public from(t: Wallet): void {
     this.id = t.id;
+    this.network = t.network;
     this.address = t.address;
     this.createdAt = t.createdAt;
   }
@@ -23,6 +27,7 @@ export class WalletModel {
   public to(): Wallet {
     return {
       id: this.id,
+      network: this.network,
       address: this.address,
       createdAt: this.createdAt,
     };
