+++ src/pkg/database/type.ts
@@ -1,4 +1,4 @@
-import { Transaction, TransactionType, Wallet } from "../../sychronizer";
+import { Network, Transaction, TransactionType, Wallet } from "../../synchronizer";
 
 export const ERROR_MSG_NOT_INITIALIZED = "db are not initalized";
 
@@ -6,10 +6,10 @@ export interface IDatabase {
   connect(): Promise<void>;
   close(): Promise<void>;
   insertTransactions(t: Transaction[]): Promise<Transaction[]>;
-  depositsByWalletContractAddress(wallet: string, contract: string): Promise<Transaction[]>;
-  withdrawsByWalletContractAddress(wallet: string, contract: string): Promise<Transaction[]>;
-  listWallets(): Promise<string[]>;
-  login(wallet: string): Promise<Wallet>;
-  listTransactionsByHashes(hashes: string[]): Promise<Transaction[]>;
-  getLastTransactionByType(type: TransactionType): Promise<Transaction>;
+  listDeposits(network: Network, wallet: string, contract: string): Promise<Transaction[]>;
+  listWithdraws(network: Network, wallet: string, contract: string): Promise<Transaction[]>;
+  listWallets(network: Network): Promise<string[]>;
+  login(network: Network, wallet: string): Promise<Wallet>;
+  listTransactionsByHashes(network: Network, hashes: string[]): Promise<Transaction[]>;
+  getLastTransactionByType(network: Network, type: TransactionType): Promise<Transaction>;
 }
